Separate Global and InitialData
